use-cases/03-ai-devops/requirements.txt
anthropic>=0.84.0
fastapi>=0.135.0
uvicorn>=0.41.0
pydantic>=2.12.0
python-dotenv>=1.2.0
httpx>=0.28.0
psutil>=6.0.0
